Version 0.2.3 - Continuous Connection Step Ticker
Update: Updated to PalBleAndroid v0.5.6

Update: Using the new version of ActivatorMicroService, the inherited ContinuousService now displays a running total of step count, time in current posture, and time stepping.
Upon connection these values are fetched from the Activator Micro and stored local. As step events an posture changes are detected, these local values are updated. Each 15-seconds, when a new summary is produced on the Activator Micro, it notified the app which fetched the summary and updates/corrects the local copy. Stepping time is only updated every 15 seconds, however, a local approximation could be implemented based upon stepping events, if a more real-time figure is required.

Fixed: In continuous connection mode disconnects were being mistakenly captured by the error processor. This resulted in reconnections being inconsistent and limited. This has been corrected, so disconnects are now handled by the ActivatorMicroService, and the library will automatically, and continuously attempt to reestablish the connection.