Version 0.2.10 - DFU Bug Fix
Update: Updated to PalBleAndroid v0.6.1

Fixed: Only the first stage of DFU was being enabled. This has now been fixed so that the second stage of DFU is correctly triggered and the results feedback to the calling activity.

Note: Version 0.6.* of the library represents the first fully unified iteration of the library.