Version 0.2.11 - Epoch List
Update: Updated to PalBleAndroid v0.6.2

Added: ActivatorMicroEpochs now has a method getList. This returns a list ActivatorMicroEpoch. These contain the summary data in addition to start and stop (and duration) timestamps. The list stores the most recent time stamp first and then on a reverse-temporal order. All epochs should be 15 seconds in duration with the exception of the first epoch which may be partically filled at time of download.